# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MarketMitra API",
    "version": "1.0.0",
    "description": "Public + session-scoped HTTP API for the MarketMitra dashboard. For automated / agent access to public market data, prefer the MCP server at /api/mcp (see /llms.txt and docs/api-surface.md). Session-scoped endpoints authenticate via the Clerk session cookie in hosted mode; in self-host mode they resolve to a single local user with no login."
  },
  "servers": [
    {
      "url": "https://marketmitra-v2.vercel.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Clerk session cookie (hosted mode). Sent automatically by the browser; the API explorer uses your current session."
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET bearer token."
      }
    },
...
  },
  "paths": {
...
    "/api/holdings": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "List holdings",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The user's holdings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          }
        }
      },
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Add a holding",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "symbol",
                  "quantity",
                  "avgPrice"
                ],
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  },
                  "avgPrice": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  }
                }
              },
              "example": {
                "symbol": "RELIANCE",
                "quantity": 10,
                "avgPrice": 2450.5
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created."
          },
          "401": {
            "description": "Not authenticated."
          },
          "422": {
            "description": "Invalid input."
          }
        }
      }
    },
...
    "/api/portfolio-import/confirm": {
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Commit a reviewed portfolio import (ADR 0022)",
        "description": "Writes the user's explicitly approved subset of a POST /api/portfolio-import/extract response — the only route that actually creates/updates holdings from a Mitra-driven import. No AI involvement; a plain create-or-update per item. Never called automatically — reached only by an explicit approve action in the UI.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "changes"
                ],
                "properties": {
                  "changes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "matchedSymbol",
                        "action",
                        "existingHoldingId",
                        "after"
                      ],
                      "properties": {
                        "matchedSymbol": {
                          "type": "string"
                        },
                        "action": {
                          "type": "string",
                          "enum": [
                            "create",
                            "update"
                          ]
                        },
                        "existingHoldingId": {
                          "type": "string",
                          "nullable": true
                        },
                        "after": {
                          "type": "object",
                          "properties": {
                            "quantity": {
                              "type": "number"
                            },
                            "avgPrice": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ success: true, data: { succeeded, failedSymbols } }"
          },
          "401": {
            "description": "Not authenticated."
          },
          "422": {
            "description": "Invalid request."
          },
          "429": {
            "description": "Rate limited."
          }
        }
      }
    },
...
    "/api/settings/guardrails": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get the auto-guardrail alerts setting",
        "description": "Whether adding a holding automatically creates a default trailing-stop + cumulative-drawdown alert pair for it (ADR 0030). Defaults to true.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The current setting.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Turn auto-guardrail alerts on or off",
        "description": "Controls whether newly added holdings get a default trailing-stop + cumulative-drawdown alert pair (ADR 0030). Does not touch alerts already created.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "enabled"
                ],
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated setting.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "422": {
            "description": "Invalid request body."
          }
        }
      }
    },
...
  }
}